Speed Min removed and uint16_t return changed to int32_t in all motor files

+++ modules/motor/motor.h
@@ -11,7 +11,7 @@
 #include "motor_timer.h"
 
 void configure_gpio(void);
-uint16_t motor_setSpeed(char cMotor, int16_t iSpeed);
+int32_t motor_setSpeed(char cMotor, int32_t iSpeed);
 //uint16_t motor_setDirection(char cMotor,char cDir);
 //uint16_t motor_setSpeedDir(char cMotor, char cDir, int iSpeed);
 void motor_stop(char cMotor);
